fix: prevent flicker
pull out asset list logic
split up into to and from token lists
limit based on number of unowned tokens
